+++ src/test/java/com/alibaba/json/bvt/serializer/filters/PropertyFilterTest2.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import junit.framework.TestCase;
