+++ src/test/java/com/alibaba/json/bvt/serializer/filters/PropertyFilter_byte.java
@@ -1,18 +1,16 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.serializer.filters;
 
 import java.util.HashMap;
 import java.util.Map;
 
+import org.junit.Assert;
 import junit.framework.TestCase;
 
-import org.junit.Assert;
-
-import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.PropertyFilter;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 
-public class PropertyFilterTest extends TestCase {
+public class PropertyFilter_byte extends TestCase {
 
     public void test_0() throws Exception {
         PropertyFilter filter = new PropertyFilter() {
@@ -32,23 +30,13 @@ public class PropertyFilterTest extends TestCase {
         String text = out.toString();
         Assert.assertEquals("{}", text);
     }
-    
-    public void test_toJSONString() throws Exception {
-        PropertyFilter filter = new PropertyFilter() {
-
-            public boolean apply(Object source, String name, Object value) {
-                return false;
-            }
-        };
-
-        Assert.assertEquals("{}", JSON.toJSONString(new A(), filter));
-    }
 
     public void test_1() throws Exception {
         PropertyFilter filter = new PropertyFilter() {
 
             public boolean apply(Object source, String name, Object value) {
                 if ("id".equals(name)) {
+                    Assert.assertTrue(value instanceof Byte);
                     return true;
                 }
                 return false;
@@ -138,14 +126,14 @@ public class PropertyFilterTest extends TestCase {
 
     public static class A {
 
-        private int    id;
+        private byte    id;
         private String name;
 
-        public int getId() {
+        public byte getId() {
             return id;
         }
 
-        public void setId(int id) {
+        public void setId(byte id) {
             this.id = id;
         }
