+++ src/test/java/com/alibaba/json/bvt/serializer/filters/ValueFilterTest_field.java
@@ -11,8 +11,7 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.SerializeWriter;
 import com.alibaba.fastjson.serializer.ValueFilter;
-import com.alibaba.json.bvt.serializer.PropertyFilterTest;
-import com.alibaba.json.bvt.serializer.PropertyFilterTest.A;
+import com.alibaba.json.bvt.serializer.filters.PropertyFilterTest.A;
 
 public class ValueFilterTest_field extends TestCase {
